+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/composable/actions/ComposableSend.kt
@@ -1,7 +1,6 @@
 package com.flipperdevices.keyscreen.impl.composable.actions
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -13,11 +12,13 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.flipperdevices.core.ui.ktx.onHoldPress
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
@@ -27,7 +28,7 @@ import com.flipperdevices.keyscreen.impl.model.FlipperDeviceState
 import com.flipperdevices.keyscreen.impl.viewmodel.FlipperDeviceViewModel
 
 @Composable
-fun ComposableEmulate(
+fun ComposableSend(
     modifier: Modifier = Modifier,
     onClick: (Boolean) -> Unit = {}
 ) {
@@ -37,43 +38,49 @@ fun ComposableEmulate(
 
     if (enabled) {
         ComposableActionDisable(
-            textId = R.string.keyscreen_emulate,
-            iconId = DesignSystem.drawable.ic_emulate
+            textId = R.string.keyscreen_send,
+            iconId = DesignSystem.drawable.ic_send
         )
-    } else ComposableEmulateInternal(
-        modifier = modifier,
-        onClick = onClick
-    )
+    } else {
+        ComposableSendInternal(
+            modifier = modifier,
+            onClick = onClick
+        )
+    }
 }
 
 @Composable
-private fun ComposableEmulateInternal(
+private fun ComposableSendInternal(
     modifier: Modifier = Modifier,
     onClick: ((Boolean) -> Unit) = {}
 ) {
     var isAction by remember { mutableStateOf(false) }
-    val textId = if (isAction) R.string.keyscreen_emulating else R.string.keyscreen_emulate
+    val scale = animateFloatAsState(if (isAction) 1.06f else 1f)
 
-    val modifierAction = modifier.clickable(
-        interactionSource = remember { MutableInteractionSource() },
-        indication = null,
-        onClick = {
-            isAction = !isAction
-            onClick.invoke(isAction)
-        }
-    )
+    val modifierAction = modifier
+        .scale(scale.value)
+        .onHoldPress(
+            onLongPressStart = {
+                isAction = true
+                onClick.invoke(isAction)
+            },
+            onLongPressEnd = {
+                isAction = false
+                onClick.invoke(isAction)
+            }
+        )
 
     ComposableActionFlipper(
         modifier = modifierAction,
-        color = LocalPallet.current.accentSecond,
-        textId = textId,
-        iconId = DesignSystem.drawable.ic_emulate,
+        color = LocalPallet.current.accent,
+        textId = R.string.keyscreen_send,
+        iconId = DesignSystem.drawable.ic_send,
         isAction = isAction
     ) {
-        if (isAction) {
+        if (!isAction) {
             Text(
                 modifier = Modifier.padding(top = 4.dp),
-                text = stringResource(id = R.string.keyscreen_emulating_desc),
+                text = stringResource(id = R.string.keyscreen_sending_desc),
                 style = LocalTypography.current.subtitleM12,
                 color = LocalPallet.current.text20,
                 textAlign = TextAlign.Center
@@ -87,12 +94,10 @@ private fun ComposableEmulateInternal(
     showBackground = true
 )
 @Composable
-private fun ComposableEmulatePreview() {
+private fun ComposableSendPreview() {
     FlipperThemeInternal {
-        Column(
-            modifier = Modifier.padding(horizontal = 24.dp).fillMaxSize()
-        ) {
-            ComposableEmulateInternal()
+        Column(modifier = Modifier.padding(horizontal = 24.dp).fillMaxSize()) {
+            ComposableSendInternal()
         }
     }
 }